refac: move mainpage block into separate files
